Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/hr-time/window-worker-timeOrigin.window-expected.txt


FAIL timeOrigin and now() should be correctly ordered between window and worker assert_less_than: Document timeOrigin must be earlier than worker timeOrigin expected a number but got a "undefined"
